+++ packages/terra-date-picker/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 ### Added
 * Added documentation on required peerDependencies
 
+### Changed
+* Updated to interface with react-intl's Intl context rather than old React Context API
+
 4.6.0 - (July 10, 2019)
 ------------------
 ### Changed